Move StringToMoney and MoneyToString to Currency unit
