+++ src/openrct2-ui/windows/Ride.cpp
@@ -40,9 +40,9 @@
 #include <openrct2/core/String.hpp>
 #include <openrct2/entity/EntityList.h>
 #include <openrct2/entity/Staff.h>
+#include <openrct2/localisation/Currency.h>
 #include <openrct2/localisation/Formatter.h>
 #include <openrct2/localisation/Localisation.Date.h>
-#include <openrct2/localisation/Localisation.h>
 #include <openrct2/localisation/LocalisationService.h>
 #include <openrct2/network/network.h>
 #include <openrct2/object/MusicObject.h>
